add: helper classes for connecting, creating and writing to DB
Added different helper classes for the database. 'Database' dataclass
for instantiating the connection to the database, as well as
'DatabaseWriter' for instantiating a writing object enabling writing to
the desire collection (equivalent to a table in normal relational DB).
Other calls to function creating the desired collections if not done
already.